refactor(auth): move AuthCard into auth feature
