+++ frontend/src/features/auth/components/AuthCard.tsx
