+++ frontend/src/pages/Register.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react"
 import { createUser } from "../lib/api"
 import { useNavigate, Link } from "react-router-dom"
-import { AuthCard } from "../components/AuthCard";
+import { AuthCard } from "../features/auth/components/AuthCard";
 import { FormField } from "../components/FormField";
 import { useApiError } from "../hooks/useApiError";
